authorized systemregister tests
